Add support for Libcamera and GStreamer

+++ backend/app/util/proc.py
@@ -1,11 +1,11 @@
 import sys
-from typing import TYPE_CHECKING, List
+from typing import TYPE_CHECKING, List, Optional
 
 if TYPE_CHECKING:
     import multiprocessing as mp
 
 
-def terminate_processes(processes: List["mp.Process"], allow_exit=False):
+def terminate_processes(processes: List[Optional["mp.Process"]], allow_exit=False):
     """
     Gracefully terminates a list of multiprocessing processes.
 
@@ -20,15 +20,37 @@ def terminate_processes(processes: List["mp.Process"], allow_exit=False):
     RESET = "\033[0m"
 
     for process in processes:
+        if process is None:
+            continue
         try:
             print(f"{BOLD}{COLOR_YELLOW}Terminating process {process.name}.{RESET}")
             process.terminate()
-            print(f"{BOLD}{COLOR_YELLOW}Joining process {process.name}.{RESET}")
-            process.join()
+
         except Exception:
             pass
 
     for process in processes:
+        if process is None:
+            continue
+        try:
+            print(f"{BOLD}{COLOR_YELLOW}Joining process {process.name}.{RESET}")
+            process.join(10)
+            print(f"{BOLD}{COLOR_YELLOW}Joined process {process.name}.{RESET}")
+
+            if process.is_alive():
+                print(
+                    f"{BOLD}{COLOR_YELLOW}Process {process.name} still alive. Terminating. {RESET}"
+                )
+                process.terminate()
+                print(f"{BOLD}{COLOR_YELLOW}Joining {process.name}. {RESET}")
+                process.join(5)
+
+        except Exception:
+            pass
+
+    for process in processes:
+        if process is None:
+            continue
         try:
             print(f"{BOLD}{COLOR_YELLOW}Closing process {process.name}.{RESET}")
             process.close()
